# template/public/apiDocsV1/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "demo-node",
    "description": "This is a sample server Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.",
    "contact": {
      "email": ""
    },
    "version": "1.0.0",
    "termsOfService": "http://swagger.io/terms/",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath": "/api/v1",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json",
    "application/x-www-form-urlencoded",
    "multipart/form-data"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/user/login": {
      "post": {
        "deprecated": false,
        "tags": [
          "User"
        ],
        "description": "Login",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "formData",
            "description": "email",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "password",
            "required": true,
            "type": "string"
          },
          {
            "name": "deviceID",
            "in": "formData",
            "description": "deviceID",
            "required": false,
            "type": "string"
          },
          {
            "name": "deviceToken",
            "in": "formData",
            "description": "deviceToken",
            "required": false,
            "type": "string"
          },
          {
            "name": "deviceTypeID",
            "in": "formData",
            "description": "1= IOS, 2=Android , 3=web",
            "required": false,
            "enum": [
              1,
              2,
              3
            ],
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": [
              "Username: ",
              "Password: "
            ]
          }
        ]
      }
    },
...
    "/user/details": {
      "post": {
        "deprecated": false,
        "tags": [
          "User"
        ],
        "description": "User Inforatation",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "formData",
            "description": "userId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": [
              "Username: ",
              "Password: "
            ]
          }
        ]
      }
    },
    "/user/info": {
      "post": {
        "deprecated": false,
        "tags": [
          "User"
        ],
        "description": "User Inforatation",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "employeeId",
            "in": "formData",
            "description": "employeeId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": [
              "Username: ",
              "Password: "
            ]
          }
        ]
      }
    },
    "/user/contactFile": {
      "post": {
        "deprecated": false,
        "tags": [
          "User"
        ],
        "description": "User contact file download",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "employeeId",
            "in": "formData",
            "description": "employeeId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": [
              "Username: ",
              "Password: "
            ]
          }
        ]
      }
    },
    "/user/forgot": {
      "post": {
        "deprecated": false,
        "tags": [
          "User"
        ],
        "description": "User forgot",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "formData",
            "description": "email",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": [
              "Username: ",
              "Password: "
            ]
          }
        ]
      }
    },
    "/user/checkForgotToken": {
      "post": {
        "deprecated": false,
        "tags": [
          "User"
        ],
        "description": "Forgot password token check",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "description": "accessToken",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": [
              "Username: ",
              "Password: "
            ]
          }
        ]
      }
    },
    "/user/reset": {
      "post": {
        "deprecated": false,
        "tags": [
          "User"
        ],
        "description": "User reset password",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "description": "accessToken",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "password",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": [
              "Username: ",
              "Password: "
            ]
          }
        ]
      }
    },
    "/user/changePassword": {
      "post": {
        "deprecated": false,
        "tags": [
          "User"
        ],
        "description": "User change password",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "description": "accessToken",
            "required": true,
            "type": "string"
          },
          {
            "name": "oldPassword",
            "in": "formData",
            "description": "oldPassword",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "password",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": [
              "Username: ",
              "Password: "
            ]
          }
        ]
      }
    },
    "/user/profile/edit": {
      "post": {
        "deprecated": false,
        "tags": [
          "User"
        ],
        "description": "User edit",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "description": "accessToken",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "name",
            "required": false,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "description": "email",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "password",
            "required": false,
            "type": "string"
          },
          {
            "name": "userType",
            "in": "formData",
            "description": "1- user , 2- company",
            "required": false,
            "type": "string"
          },
          {
            "name": "gender",
            "in": "formData",
            "description": "gender",
            "required": false,
            "type": "string"
          },
          {
            "name": "dob",
            "in": "formData",
            "description": "dob",
            "required": false,
            "type": "string"
          },
          {
            "name": "designation",
            "in": "formData",
            "description": "designation",
            "required": false,
            "type": "string"
          },
          {
            "name": "companyName",
            "in": "formData",
            "description": "companyName",
            "required": false,
            "type": "string"
          },
          {
            "name": "aboutUs",
            "in": "formData",
            "description": "aboutUs",
            "required": false,
            "type": "string"
          },
          {
            "name": "profileImage",
            "in": "formData",
            "description": "profileImage",
            "required": false,
            "type": "file"
          },
          {
            "name": "workEmail",
            "in": "formData",
            "description": "workEmail",
            "required": false,
            "type": "string"
          },
          {
            "name": "phoneNo",
            "in": "formData",
            "description": "phoneNo",
            "required": false,
            "type": "string"
          },
          {
            "name": "whatsApp",
            "in": "formData",
            "description": "whatsApp",
            "required": false,
            "type": "string"
          },
          {
            "name": "skype",
            "in": "formData",
            "description": "skype",
            "required": false,
            "type": "string"
          },
          {
            "name": "website",
            "in": "formData",
            "description": "website",
            "required": false,
            "type": "string"
          },
          {
            "name": "instagram",
            "in": "formData",
            "description": "instagram",
            "required": false,
            "type": "string"
          },
          {
            "name": "facebook",
            "in": "formData",
            "description": "facebook",
            "required": false,
            "type": "string"
          },
          {
            "name": "linkedin",
            "in": "formData",
            "description": "linkedin",
            "required": false,
            "type": "string"
          },
          {
            "name": "twitter",
            "in": "formData",
            "description": "twitter",
            "required": false,
            "type": "string"
          },
          {
            "name": "googleMap",
            "in": "formData",
            "description": "googleMap",
            "required": false,
            "type": "string"
          },
          {
            "name": "hangouts",
            "in": "formData",
            "description": "hangouts",
            "required": false,
            "type": "string"
          },
          {
            "name": "youtube",
            "in": "formData",
            "description": "youtube",
            "required": false,
            "type": "string"
          },
          {
            "name": "snapchat",
            "in": "formData",
            "description": "snapchat",
            "required": false,
            "type": "string"
          },
          {
            "name": "pinterest",
            "in": "formData",
            "description": "pinterest",
            "required": false,
            "type": "string"
          },
          {
            "name": "tiktok",
            "in": "formData",
            "description": "tiktok",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": [
              "Username: ",
              "Password: "
            ]
          }
        ]
      }
    },
    "/user/create": {
      "post": {
        "deprecated": false,
        "tags": [
          "Admin"
        ],
        "description": "User Add",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "description": "accessToken",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "name",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "description": "email",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "password",
            "required": false,
            "type": "string"
          },
          {
            "name": "userType",
            "in": "formData",
            "description": "1- user , 2- company",
            "required": false,
            "type": "string"
          },
          {
            "name": "gender",
            "in": "formData",
            "description": "gender",
            "required": false,
            "type": "string"
          },
          {
            "name": "dob",
            "in": "formData",
            "description": "dob",
            "required": false,
            "type": "string"
          },
          {
            "name": "designation",
            "in": "formData",
            "description": "designation",
            "required": false,
            "type": "string"
          },
          {
            "name": "companyName",
            "in": "formData",
            "description": "companyName",
            "required": false,
            "type": "string"
          },
          {
            "name": "aboutUs",
            "in": "formData",
            "description": "aboutUs",
            "required": false,
            "type": "string"
          },
          {
            "name": "profileImage",
            "in": "formData",
            "description": "profileImage",
            "required": false,
            "type": "file"
          },
          {
            "name": "workEmail",
            "in": "formData",
            "description": "workEmail",
            "required": false,
            "type": "string"
          },
          {
            "name": "phoneNo",
            "in": "formData",
            "description": "phoneNo",
            "required": false,
            "type": "string"
          },
          {
            "name": "whatsApp",
            "in": "formData",
            "description": "whatsApp",
            "required": false,
            "type": "string"
          },
          {
            "name": "skype",
            "in": "formData",
            "description": "skype",
            "required": false,
            "type": "string"
          },
          {
            "name": "website",
            "in": "formData",
            "description": "website",
            "required": false,
            "type": "string"
          },
          {
            "name": "instagram",
            "in": "formData",
            "description": "instagram",
            "required": false,
            "type": "string"
          },
          {
            "name": "facebook",
            "in": "formData",
            "description": "facebook",
            "required": false,
            "type": "string"
          },
          {
            "name": "linkedin",
            "in": "formData",
            "description": "linkedin",
            "required": false,
            "type": "string"
          },
          {
            "name": "twitter",
            "in": "formData",
            "description": "twitter",
            "required": false,
            "type": "string"
          },
          {
            "name": "googleMap",
            "in": "formData",
            "description": "googleMap",
            "required": false,
            "type": "string"
          },
          {
            "name": "hangouts",
            "in": "formData",
            "description": "hangouts",
            "required": false,
            "type": "string"
          },
          {
            "name": "youtube",
            "in": "formData",
            "description": "youtube",
            "required": false,
            "type": "string"
          },
          {
            "name": "snapchat",
            "in": "formData",
            "description": "snapchat",
            "required": false,
            "type": "string"
          },
          {
            "name": "pinterest",
            "in": "formData",
            "description": "pinterest",
            "required": false,
            "type": "string"
          },
          {
            "name": "tiktok",
            "in": "formData",
            "description": "tiktok",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": [
              "Username: ",
              "Password: "
            ]
          }
        ]
      }
    },
...
    "/user/list": {
      "post": {
        "deprecated": false,
        "tags": [
          "Admin"
        ],
        "description": "User List",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "description": "accessToken",
            "required": true,
            "type": "string"
          },
          {
            "name": "userType",
            "in": "formData",
            "description": "userType",
            "required": true,
            "type": "string"
          },
          {
            "name": "companyId",
            "in": "formData",
            "description": "companyId",
            "required": false,
            "type": "string"
          },
          {
            "name": "search",
            "in": "formData",
            "description": "search",
            "required": false,
            "type": "string"
          },
          {
            "name": "pageNo",
            "in": "formData",
            "description": "pageNo",
            "required": false,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "formData",
            "description": "limit",
            "required": false,
            "type": "string"
          },
          {
            "name": "sortField",
            "in": "formData",
            "description": "sortField",
            "required": false,
            "type": "string"
          },
          {
            "name": "sortType",
            "in": "formData",
            "description": "sortType",
            "required": false,
            "type": "string"
          },
          {
            "name": "status",
            "in": "formData",
            "description": "status",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": [
              "Username: ",
              "Password: "
            ]
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "basicAuth": {
      "type": "basic"
    }
  }
}